Fixed some bugs;

+++ Projeto_A.X/adc.h
@@ -8,7 +8,7 @@ typedef enum{AN0=0,AN1,AN2,AN3,AN4,AN5,AN6,AN7,AN8,AN9,AN10,AN11,AN12,AN13,AN14,
 void ADC_init(void);
 int ADC_input(uint16_t a);
 void ADC_enable(void);
-float ADC_read(void);
+uint16_t ADC_read(void);
 uint8_t ADC_IF(void);
 void ADC_start(void);
 #endif
